bugfix: fix the cli args for normal startup of lsp to be consistent with client of vscode
