+++ kclvm/tools/src/LSP/Cargo.toml
@@ -11,7 +11,7 @@ indexmap = "1.0"
 chumsky = "0.9.0"
 env_logger = "0.10.0"
 ropey = "1.3.2"
-tokio = { version = "1.17.0", features = ["full"] }
+tokio = { version = "1.0.0", features = ["full"] }
 tokio-test = "0.4.2"
 serde = { version = "1.0", features = ["derive"] }
 dashmap = "5.1.0"
