+++ kclvm/tools/src/LSP/src/util.rs
@@ -13,8 +13,8 @@ use kclvm_error::Position as KCLPos;
 use kclvm_parser::entry::get_dir_files;
 use kclvm_parser::{load_program, ParseSession};
 use kclvm_sema::resolver::resolve_program_with_opts;
+use kclvm_sema::resolver::scope::ProgramScope;
 use kclvm_sema::resolver::scope::Scope;
-use kclvm_sema::resolver::{resolve_program, scope::ProgramScope};
 use kclvm_utils::pkgpath::rm_external_pkg_name;
 use lsp_types::Url;
 use parking_lot::{RwLock, RwLockReadGuard};
